pull out stats into tilingstats module
