+++ doc/source/pyshepseg_tilingstats.rst
@@ -1,0 +1,9 @@
+tilingstats
+===========
+.. automodule:: pyshepseg.tilingstats
+   :members:
+   :undoc-members:
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
